move useInfiniteQuery to Hook
