+++ src/hooks/useListing.js
@@ -1,32 +1,66 @@
-import { useQuery } from 'react-query';
+import {useInfiniteQuery, useQuery} from 'react-query';
 import axios from 'axios';
-import {apiEndpoint} from "../constants";
+import {apiEndpoint, marketContractAddress, zoombiesContractAddress} from "../constants";
+import {getTokenSymbol} from "../utils/auction";
 
-const getListings = async ( filters ) => {
+export const getAuctionListings = async (marketContract, zoombiesContract, filters, sorting, page) => {
+  // console.log({filters, sorting})
+
+  const getSortType = () => {
+    // switch(sorting.field) {
+    //   case 'auctionEnd':
+    //     return 'END_TIME'
+    //   case 'minPrice':
+    //     return 'MIN_PRICE'
+    //   case 'highestBid':
+    //     return 'HIGHEST_BID'
+    //   case '':
+    //     return null
+    //   default:
+    //     throw new Error(`Unhandled sort type: ${sorting.field}`)
+    // }
+    return sorting.field
+  }
 
   const params = new URLSearchParams({
     cardOrigin: filters.cardType,
     saleToken: filters.token,
     cardRarity: filters.rarity,
     search: filters.keyword,
-    sortBy: filters.sortField
+    sortBy: getSortType() ?? '',
+    offset: page * 5,
+    limit: '5'
   })
 
-  const response = await axios.get(`${apiEndpoint}/listing/${params.toString()}`);
+  const listings = await axios.get(`https://api.zoombies.world/listings?${params.toString()}`)
+  // const listings = await axios.get(`http://localhost:3001/listings?${params.toString()}`)
 
-  if (response.status !== 200) {
-      console.error(response.statusText);
-      throw new Error("Failed to get profile.");
+  const ar2 = listings.data.listings
+  return {
+    data: ar2.map((listing) => ({
+      ...listing,
+      id: listing._id,
+      currency: getTokenSymbol(listing.saleToken),
+    })),
+    nextPage: page + 1, totalCount: parseInt(listings.data.count)
   }
-
-  return response.data;
 };
 
-export const useFetchListingQuery = ( filters ) =>
-  useQuery({
-    queryKey: ['fetchListing', JSON.stringify(filters)],
-    queryFn: () => getListings(filters),
-    ...{
-      refetchOnWindowFocus: false,
-    },
-  });
+export const useFetchListingQuery = ( filters, sortBy, callback ) => {
+  return useInfiniteQuery(
+    'listings',
+    async ({ pageParam = 0 }) => {
+      const res = await getAuctionListings(marketContractAddress, zoombiesContractAddress, filters, sortBy, pageParam)
+      if (callback) {
+        callback(res.totalCount)
+      }
+      return res
+
+    },{
+      getNextPageParam: (lastPage, pages) => {
+        if (lastPage.nextPage < (lastPage.totalCount / 5)) return lastPage.nextPage;
+        return undefined;
+      }
+    }
+  )
+}
